Adicionando as restrições do client

+++ src/controllers/DeviceController.js
@@ -1,33 +1,31 @@
 const Device = require('../models/Device');
 
-module.exports = {
-     
-    async index(req, res) {
-        const { user_id } = req.headers;
+let contador = 0;
+var mud = 1586185682411;
+var hoje = new Date();
 
+module.exports = {    
+     
+    async index(req, res) {        
+        const { user_id } = req.headers;
         const device = await Device.find({ id: user_id })
 
         return res.json(device)
     },
 
-    async store(req, res) {
-        let contador = 3;       
-        
-
-        if(contador < 3){            
-
-            const { phone } = req.body;
+    async store(req, res) {                   
+        if(contador < 3){
+            const { nome } = req.body;
             const { user_id } = req.headers;
-            const { createdAt } = req.body;
-            
+            const { modelo } = req.body;
 
             let device = await Device.create({
                 user: user_id,
-                createdAt,
-                phone,
+                modelo,
+                nome,
             });
 
-        
+        contador++
         return res.json(device);
         }else{
             return res.send()
@@ -35,15 +33,27 @@ module.exports = {
     },
 
     async update(req, res) {
-        const device = await Device.findOneAndUpdate(req.params.id, req.body, { new:true })
+        if(hoje > mud) {            
+            var novaData = adicionarDiasData(30);
+            mud = novaData;
+
+            const device = await Device.findOneAndUpdate(req.params.id, req.body, { new:true })
 
         return res.json(device);
+        }else{
+            return res.send({ error: `Você poderá fazer uma nova mudança na data: ${mud}` })
+        }
     },
 
     async destroy(req, res){
-        await Device.findByIdAndRemove(req.params.id);
+        if(hoje > mud) {
+            await Device.findByIdAndRemove(req.params.id);
 
-        return res.send();
+            contador--
+            return res.send();
+        }else{
+            return res.send({ error: `Você poderá fazer uma nova mudança ${mud}` })
+        }
     }
